+++ sdk/storage/Azure.Storage.Files.Shares/src/Models/StorageError.cs
@@ -48,6 +48,11 @@ namespace Azure.Storage.Files.Shares.Models
         /// A <see cref="RequestFailedException"/>.
         /// </returns>
         public Exception CreateException(ClientDiagnostics clientDiagnostics, Azure.Response response)
-            => clientDiagnostics.CreateRequestFailedExceptionWithContent(response, message: Message, content: null, response.GetErrorCode(Code));
+            => clientDiagnostics.CreateRequestFailedExceptionWithContent(
+                response,
+                message: Message,
+                content: null,
+                response.GetErrorCode(Code),
+                AdditionalInformation);
     }
 }