+++ sdk/storage/Azure.Storage.Blobs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 12.7.0-preview.2 (Unreleased)
 - Fixed bug where BobContainerClient.SetAccessPolicy() would throw an exception if signed identifier permissions were not in the correct order.
 - Added seekability to BaseBlobClient.OpenRead().
+- Added additional info to exception messages.
 
 ## 12.7.0-preview.1 (2020-09-30)
 - Added support for service version 2020-02-10.